Allow cache_bench/db_bench to use a custom secondary cache
