+++ src.mk
@@ -320,6 +320,9 @@ BENCH_LIB_SOURCES =                                             \
   tools/db_bench_tool.cc                                        \
   tools/simulated_hybrid_file_system.cc                         \
 
+CACHE_BENCH_LIB_SOURCES =					\
+  cache/cache_bench_tool.cc                                     \
+
 STRESS_LIB_SOURCES =                                            \
   db_stress_tool/batched_ops_stress.cc                         \
   db_stress_tool/cf_consistency_stress.cc                      \
